Scale down test should now function

+++ src/acceptance/app/recurring_schedule_policy_test.go
@@ -3,6 +3,8 @@ package app_test
 import (
 	"acceptance/config"
 	. "acceptance/helpers"
+	"fmt"
+	"os"
 
 	"github.com/cloudfoundry-incubator/cf-test-helpers/cf"
 	"github.com/cloudfoundry-incubator/cf-test-helpers/generator"
@@ -48,8 +50,12 @@ var _ = Describe("AutoScaler recurring schedule policy", func() {
 	})
 
 	AfterEach(func() {
-		DeletePolicy(appName, appGUID)
-		Expect(cf.Cf("delete", appName, "-f", "-r").Wait(cfg.DefaultTimeoutDuration())).To(Exit(0))
+		if os.Getenv("SKIP_TEARDOWN") == "true" {
+			fmt.Println("Skipping Teardown...")
+		} else {
+			DeletePolicy(appName, appGUID)
+			Expect(cf.Cf("delete", appName, "-f", "-r").Wait(cfg.DefaultTimeoutDuration())).To(Exit(0))
+		}
 	})
 
 	Context("when scaling by recurring schedule", func() {
